requirements.txt
numpy
pandas
seaborn
matplotlib
ipython
scrapy
xlrd
openpyxl
unidecode
unicode-slugify
